+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 * sdk: document and rename `Client::metadata` to `Client::fetch_metadata` ([Janek])
 * sdk: update `Client::shutdown` method fingerprint ([Yuki Kishimoto])
 * sdk: deprecate `Client::add_relay_with_opts` and `Client::add_relays` ([Yuki Kishimoto])
+* sdk: deprecate `RelayPool::send_msg` and `RelayPool::batch_msg` ([Yuki Kishimoto])
 * ffi(nostr): impl `Display` for `Coordinate` ([Yuki Kishimoto])
 * ffi(sdk): change `Connection::embedded_tor` fingerprint for `android` and `ios` targets ([Yuki Kishimoto])